requirements.txt
acres==0.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dill==0.4.0
docling==2.31.0
docling-core==2.29.0
docling-ibm-models==3.4.2
docling-parse==4.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
isodate==0.6.1
Jinja2==3.1.6
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
latex2mathml==3.78.0
lazy_loader==0.4
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.7.1
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
narwhals==1.38.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==1.26.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
pluggy==1.5.0
protobuf==6.30.2
prov==2.0.1
puremagic==1.29
pyarrow==20.0.0
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
Pygments==2.19.1
pylatexenc==2.10
PyMuPDF==1.26.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semchunk==2.2.2
setuptools==80.3.1
shapely==2.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.45.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.3.30
tokenizers==0.21.1
toml==0.10.2
torch==2.2.0+cpu
torchvision==0.17.0+cpu
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
XlsxWriter==3.2.3
